Split right-panel migration helpers
